refactor(keyboard): split keycap appearance theme tokens into separate file
